Refactor rustdoc processing
Also avoid a panic when cargo-sync-rdme cannot resolve some
valid intra-doc links in rustdoc output.
